pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
fallback_version = "0.0"

[project]
classifiers = [
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
name = "nomad-distribution"
description = "nomad distribution template"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "nomad-lab[infrastructure]>=1.3.10",
  "setuptools",
  "nomad-schema-plugin-run>=1.0.1",
  "nomad-parser-plugins-electronic>=1.0",
  "nomad-parser-plugins-atomistic>=1.0.6",
  "ccpnc-oasis-app @ git+https://github.com/CCP-NC/ccpnc-oasis-app.git@afa087841740930af46ed97f5ce1c49de380b7f7",
  "nomad-oasis-schema-parser-plugin @ git+https://github.com/CCP-NC/nomad-oasis-schema-parser-plugin.git@9f5b1e87e82b5e6b83497c9968407b027883451f",
]


[project.urls]
repository = "https://github.com/sathya-s3/nomad-distro-template-test2"

[project.optional-dependencies]
# Add your plugins to the list below.
plugins = []
jupyter = [
  "jupyter-events",
  "voila",
  "ipyaggrid",
  "ipysheet",
  "ipydatagrid",
  "jupyter-flex",
  "pyOpenSSL",
]
gpu-action = []
cpu-action = []

dev = ["nomad-lab[infrastructure, dev]"]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]
prerelease = "explicit"
constraint-dependencies = ["flatdict>=4.1.0"]
#remove override when final version is available
override-dependencies = [
  "nomad-nmr-schema @ git+https://github.com/FAIRmat-NFDI/nomad-schema-plugin-nmr.git@f799401d8981eb9080f18e26d524d8eb7535488e",
]

[tool.setuptools.packages.find]
exclude = ["gui*", "docs*", "config*"]
